Add variable host and change http.Get request format
Introduced a new variable host in common.go to dynamically provide the RSS host URL in fetchRss function. This change allows customizing the host URL where the RSS feed is fetched from, making the function more flexible to use with different RSS hosts. Flag for the host is added in main.go to input the host while running the application.